RUN--3387 renamed legacy external app to deprecated external window
