+++ src/browser/api_protocol/api_handlers/deprecated_external_windowing_middleware.ts
